backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.4
Pillow>=10.1
psycopg2-binary>=2.9
python-decouple>=3.8
SpeechRecognition>=3.10
pydub>=0.25
dj-database-url>=2.1
